+++ CHANGELOG.md
@@ -5,6 +5,19 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- add official support of drupal 10.0
+
+### Changed
+- re-enable PHPUnit Symfony Deprecation notice
+
+### Fixed
+- fix PHPUnit deprecated prophecy integration
+- fix deprecated uncalling accessCheck relying on entity queries to check access
+- fix usage of deprecated Symfony\Component\EventDispatcher\Event in favor of Symfony\Contracts\EventDispatcher\Event
+- fix deprecation of theme classy for Drupal 10 compatibilities
+- fix tests self::modules property must be declared protected
+- fix Symfony 4.4 event dispatcher parameter order change
 
 ## [2.0.0] - 2023-01-27
 ### Fixed